requirements.txt
requests
beautifulsoup4
bs4
lxml
pandas
python-dotenv